+++ java/server/src/org/openqa/selenium/cli/CliCommand.java
@@ -17,13 +17,15 @@
 
 package org.openqa.selenium.cli;
 
+import java.io.PrintStream;
+
 public interface CliCommand {
 
   String getName();
 
   String getDescription();
 
-  Executable configure(String... args);
+  Executable configure(PrintStream out, PrintStream err, String... args);
 
   default boolean isShown() {
     return true;
@@ -31,7 +33,7 @@ public interface CliCommand {
 
   interface Executable {
 
-    void run() throws Exception;
+    void run();
   }
 
 }