+++ java/server/src/org/openqa/selenium/grid/TemplateGridCommand.java
@@ -19,6 +19,8 @@ package org.openqa.selenium.grid;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.ParameterException;
+import com.beust.jcommander.internal.DefaultConsole;
+
 import org.openqa.selenium.cli.CliCommand;
 import org.openqa.selenium.grid.config.AnnotatedConfig;
 import org.openqa.selenium.grid.config.CompoundConfig;
@@ -29,13 +31,14 @@ import org.openqa.selenium.grid.config.EnvConfig;
 import org.openqa.selenium.grid.log.LoggingOptions;
 import org.openqa.selenium.grid.server.HelpFlags;
 
+import java.io.PrintStream;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
 public abstract class TemplateGridCommand implements CliCommand {
 
   @Override
-  public final Executable configure(String... args) {
+  public final Executable configure(PrintStream out, PrintStream err, String... args) {
     Set<Object> allFlags = getFlagObjects();
 
     HelpFlags helpFlags = new HelpFlags();
@@ -47,17 +50,18 @@ public abstract class TemplateGridCommand implements CliCommand {
       .addObject(helpFlags);
     allFlags.forEach(builder::addObject);
     JCommander commander = builder.build();
+    commander.setConsole(new DefaultConsole(out));
 
     return () -> {
       try {
         commander.parse(args);
       } catch (ParameterException e) {
-        System.err.println(e.getMessage());
+        err.println(e.getMessage());
         commander.usage();
         return;
       }
 
-      if (helpFlags.displayHelp(commander, System.out)) {
+      if (helpFlags.displayHelp(commander, out)) {
         return;
       }
 
@@ -70,7 +74,7 @@ public abstract class TemplateGridCommand implements CliCommand {
 
       Config config = new CompoundConfig(allConfigs.toArray(new Config[0]));
 
-      if (helpFlags.dumpConfig(config, System.out)) {
+      if (helpFlags.dumpConfig(config, out)) {
         return;
       }
 
@@ -87,5 +91,5 @@ public abstract class TemplateGridCommand implements CliCommand {
 
   protected abstract Config getDefaultConfig();
 
-  protected abstract void execute(Config config) throws Exception;
+  protected abstract void execute(Config config);
 }