+++ java/server/src/org/openqa/selenium/grid/commands/InfoCommand.java
@@ -19,12 +19,16 @@ package org.openqa.selenium.grid.commands;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.ParameterException;
+import com.beust.jcommander.internal.DefaultConsole;
+
 import com.google.auto.service.AutoService;
 import com.google.common.io.Resources;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.PrintStream;
 import java.io.StringReader;
+import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import org.openqa.selenium.cli.CliCommand;
 import org.openqa.selenium.cli.WrappedPrintWriter;
@@ -32,7 +36,6 @@ import org.openqa.selenium.grid.server.HelpFlags;
 
 import java.io.PrintWriter;
 import java.util.Collections;
-
 
 @AutoService(CliCommand.class)
 public class InfoCommand implements CliCommand {
@@ -45,7 +48,7 @@ public class InfoCommand implements CliCommand {
     return "Prints information for commands and topics.";
   }
 
-  public Executable configure(String... args) {
+  public Executable configure(PrintStream out, PrintStream err, String... args) {
     HelpFlags help = new HelpFlags();
     InfoFlags topic = new InfoFlags();
 
@@ -54,17 +57,18 @@ public class InfoCommand implements CliCommand {
       .addObject(help)
       .addObject(topic)
       .build();
+    commander.setConsole(new DefaultConsole(out));
 
     return () -> {
       try {
         commander.parse(args);
       } catch (ParameterException e) {
-        System.err.println(e.getMessage());
+        err.println(e.getMessage());
         commander.usage();
         return;
       }
 
-      if (help.displayHelp(commander, System.out)) {
+      if (help.displayHelp(commander, out)) {
         return;
       }
 
@@ -90,13 +94,18 @@ public class InfoCommand implements CliCommand {
       }
 
       String path = getClass().getPackage().getName().replaceAll("\\.", "/") + "/" + toDisplay;
-      String content = readContent(path);
+      String content;
+      try {
+        content = readContent(path);
+      } catch (IOException e) {
+        throw new UncheckedIOException(e);
+      }
 
-      PrintWriter out = new WrappedPrintWriter(System.out, 72, 0);
+      PrintWriter outWriter = new WrappedPrintWriter(out, 72, 0);
 
-      out.printf("\n%s\n%s\n\n", title, String.join("", Collections.nCopies(title.length(), "=")));
-      out.print(content);
-      out.println("\n\n");
+      outWriter.printf("\n%s\n%s\n\n", title, String.join("", Collections.nCopies(title.length(), "=")));
+      outWriter.print(content);
+      outWriter.println("\n\n");
     };
   }