+++ java/server/src/org/openqa/selenium/grid/commands/MessageBusCommand.java
@@ -22,7 +22,9 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.openqa.selenium.BuildInfo;
 import org.openqa.selenium.cli.CliCommand;
+import org.openqa.selenium.events.Event;
 import org.openqa.selenium.events.EventBus;
+import org.openqa.selenium.events.Type;
 import org.openqa.selenium.grid.TemplateGridCommand;
 import org.openqa.selenium.grid.config.Config;
 import org.openqa.selenium.grid.config.MapConfig;
@@ -37,6 +39,8 @@ import org.openqa.selenium.remote.http.HttpResponse;
 import org.openqa.selenium.remote.http.Route;
 
 import java.util.Set;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
 import static org.openqa.selenium.json.Json.JSON_UTF_8;
@@ -65,7 +69,6 @@ public class MessageBusCommand extends TemplateGridCommand {
     return ImmutableSet.of(
       new BaseServerFlags(),
       new EventBusFlags());
-
   }
 
   @Override
@@ -85,19 +88,33 @@ public class MessageBusCommand extends TemplateGridCommand {
   }
 
   @Override
-  protected void execute(Config config) throws Exception {
+  protected void execute(Config config) {
     EventBusOptions events = new EventBusOptions(config);
-    // We need this reference to stop the bus being garbage collected. Which would be less than ideal.
     EventBus bus = events.getEventBus();
 
     BaseServerOptions serverOptions = new BaseServerOptions(config);
 
     Server<?> server = new NettyServer(
       serverOptions,
-      Route.get("/status").to(() -> req ->
-        new HttpResponse()
-          .addHeader("Content-Type", JSON_UTF_8)
-          .setContent(Contents.asJson(ImmutableMap.of("ready", true, "message", "Event bus running")))));
+      Route.get("/status").to(() -> req -> {
+        CountDownLatch latch = new CountDownLatch(1);
+
+        Type healthCheck = new Type("healthcheck");
+        bus.addListener(healthCheck, event -> latch.countDown());
+        bus.fire(new Event(healthCheck, "ping"));
+
+        try {
+          if (latch.await(5, TimeUnit.SECONDS)) {
+            return httpResponse(true, "Event bus running");
+          } else {
+            return httpResponse(false, "Event bus could not deliver a test message in 5 seconds");
+          }
+        } catch (InterruptedException e) {
+          Thread.currentThread().interrupt();
+          return httpResponse(false, "Status checking was interrupted");
+        }
+      })
+    );
     server.start();
 
     BuildInfo info = new BuildInfo();
@@ -106,10 +123,13 @@ public class MessageBusCommand extends TemplateGridCommand {
       info.getReleaseLabel(),
       info.getBuildRevision(),
       server.getUrl()));
+  }
 
-    // If we exit, the bus goes out of scope, and it's closed
-    Thread.currentThread().join();
-
-    LOG.info("Shutting down: " + bus);
+  private HttpResponse httpResponse(boolean ready, String message) {
+    return new HttpResponse()
+        .addHeader("Content-Type", JSON_UTF_8)
+        .setContent(Contents.asJson(ImmutableMap.of(
+            "ready", ready,
+            "message", message)));
   }
 }