+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -5,6 +5,8 @@
     - fileUploadDialogTimeout
     - setEdgePath
     - setEdgeChromium
+    - setScrollBehavior
+    - Supports adding multiple command-line switches using options.addArguments()
 
 ### API Changes